Refactor useRequest to solve for infinite loop bug
Using the return value of the `useRequest` hook as a dependency
for another hook formerly led to infinite loops. The workaround
to date has been to not provide the return value as a dependency
and therefore tolerate eslint warnings.

With this change, can use returned value as a conventional hook dependency.