DEV-2297 Auto-generate Donation Page Name
If name is not provided to endpoint one is auto-generated based off of
Revenue Program Name with serial integer appended. Or, Template.name if
template_pk is provided. A name is provided to endpoint it will override
auto-generation.

Note: DonationPage.name is not unique, nor unique together with
RevenueProgram. There is no technical need for appending serial int to
name.

Note: The serial int is not monotonically increasing. If previous
"autogenerated" name is deleted. That name will be reused.

Note: DonationPage based off of template will not have serial integer
appended and will not generate unique name.

Also semi-related to ticket at hand,

  - 100% coverage for apps/pages [I added expected failing test cause I don't know deal with sidebar_elements]
  - Add missing tests around Donation Page creation with/without templates and unowned RP.
  - Add comments / docstring updates (to unrelated ./pages/ files).
  - Cut "helper" file / reduce indirection.